requirements.txt
asgiref==3.6.0
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==6.6.0
Markdown==3.4.3
psycopg2==2.9.6
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.4
zipp==3.15.0
